[CCP-246] Seperated the business logic and db connection logic. Also,made changes for geeting the enry id on click of Review Button.
