+++ Conceptpower+Spring/src/main/java/edu/asu/conceptpower/app/db4o/ICommentsDBManager.java
@@ -1,0 +1,16 @@
+package edu.asu.conceptpower.app.db4o;
+
+
+import com.db4o.ObjectSet;
+
+import edu.asu.conceptpower.core.ReviewRequest;
+
+public interface ICommentsDBManager {
+
+    
+    public abstract void store(ReviewRequest reviewRequest);
+
+    ObjectSet<ReviewRequest> queryByExample(Object example);
+
+    
+}
